setting up user routers in newly created routers folder
