+++ src/index.js
@@ -11,80 +11,6 @@ app.use(express.json()); // this configures express to automatically parse JSON 
 // object so we can access it in our request handlers
 app.use(userRouter);
 
-app.post("/users", async (req, res) => {
-  // creating new instance of User
-  // console.log(req.body);
-  const user = new User(req.body);
-  try {
-    await user.save();
-    res.status(201).send(user);
-  } catch (e) {
-    res.status(400).send(e.message);
-  }
-});
-
-app.get("/users", async (req, res) => {
-  // If we live the Object blank, it fetches all of the users from database
-  try {
-    const users = await User.find({});
-    res.send(users);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
-app.get("/users/:id", async (req, res) => {
-  console.log(req.params);
-  // route parameters--> id can be named by anything else
-  const _id = req.params.id;
-  try {
-    const user = await User.findById(_id);
-    if (!user) {
-      // mongodb does not return an error if the ID does not match up in database
-      return res.status(404).send(); // 404: not found
-    }
-    res.send(user);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
-app.patch("/users/:id", async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ["name", "email", "password", "age"];
-  const isValidOperation = updates.every(update =>
-    allowedUpdates.includes(update)
-  );
-
-  if (!isValidOperation) {
-    return res.status(400).send({ error: "Invalid updates!" });
-  }
-  try {
-    const user = await User.findByIdAndUpdate(req.params.id, req.body, {
-      new: true, // to return the modified document rather than the original. defaults to false
-      runValidators: true // if true, runs update validators on this command
-    });
-    if (!user) {
-      return res.status(404).send();
-    }
-    res.send(user);
-  } catch (e) {
-    res.status(400).send(e.message);
-  }
-});
-
-app.delete("/users/:id", async (req, res) => {
-  const user = await User.findByIdAndDelete(req.params.id);
-  try {
-    if (!user) {
-      return res.status(404).send();
-    }
-    res.send(user);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
 app.post("/tasks", async (req, res) => {
   const task = new Task(req.body);
   try {
